+++ CHANGELOG.rst
@@ -8,8 +8,8 @@ Latest
 * Drop support for Python 3.6.
 * Add support for default Field values.
 * Add EnumField.
+* Support warnings in contract checks.
 * Add unmatched_ignore_imports_alerting option for each contract.
-
 
 1.2.6 (2021-09-24)
 ------------------